backend/pyproject.toml
[project]
name = "face-access-control"
version = "1.0.0"
description = "Web-based facial access control system backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "onnxruntime>=1.16.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "aiohttp>=3.9.0",      # Async HTTP client for model download
    "tqdm>=4.66.0",        # Progress bar for download
]

[tool.hatch.build.targets.wheel]
packages = ["core", "routes", "utils", "models"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
